+++ sphinx/source/changelog.rst
@@ -13,6 +13,8 @@ Changelog
   ``TraceOperand`` configuration parameters
 * Added the ``TraceHeader`` configuration parameter for
   :ref:`tap2sna.py <tap2sna-conf>` (to specify the header for a trace log file)
+* Added the ``--analyse`` option to :ref:`tapinfo.py` (for showing an analysis
+  of the tape's tones, pulse sequences and data blocks)
 
 10.1 (2026-08-14)
 -----------------